+++ pennylane_lightning/lightning_qubit/_adjoint_jacobian.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 from warnings import warn
 
 try:
-    from pennylane_lightning.lightning_qubit_ops.algorithms import (
+    from pennylane_lightning.lightning_qubit_nb.algorithms import (
         AdjointJacobianC64,
         AdjointJacobianC128,
         create_ops_listC64,